ProgettoDSBD/data_storage/requirements.txt
flask
kafka-python
prometheus_api_client
confluent_kafka
prometheus_client
mysql.connector
